feat: add new services
tags, categories services created
