+++ src/apis/post.ts
@@ -7,7 +7,9 @@ export const postRouter = Router()
 postRouter.get(
   '/posts',
   asyncHandler(async (req, res) => {
-    const posts = await PostController.getPosts()
+    const posts = (await PostController.getPosts()).sort((a, b) =>
+      a.publishedAt > b.publishedAt ? -1 : 0
+    )
     res.json(posts)
   })
 )
@@ -17,7 +19,6 @@ postRouter.get(
   asyncHandler(async (req, res) => {
     const fileName = req.params.fileName
     if (!fileName) throw new Error('No file name provided')
-
     res.json(await PostController.getPost(fileName))
   })
 )
